Windows 64-bit support

+++ gameplay/src/PhysicsVehicle.cpp
@@ -233,7 +233,7 @@ void PhysicsVehicle::setEnabled(bool enable)
 
 unsigned int PhysicsVehicle::getNumWheels() const
 {
-    return _wheels.size();
+    return (unsigned int)_wheels.size();
 }
 
 PhysicsVehicleWheel* PhysicsVehicle::getWheel(unsigned int i)
@@ -243,7 +243,7 @@ PhysicsVehicleWheel* PhysicsVehicle::getWheel(unsigned int i)
 
 void PhysicsVehicle::addWheel(PhysicsVehicleWheel* wheel)
 {
-    unsigned int i = _wheels.size();
+    unsigned i = (unsigned int)_wheels.size();
     _wheels.push_back(wheel);
     wheel->setHost(this, i);
     wheel->addToVehicle(_vehicle);
